+++ components/screens/head/index.js
@@ -2,91 +2,8 @@ import Head from "next/head";
 import Script from "next/script";
 import PropTypes from "prop-types";
 import languages from "@config/languages.json";
-import { useRouter } from "next/router";
 
-const MainHead = ({ title = null, description = null, isDesktopClient }) => {
-  const router = useRouter();
-  let style;
-
-  if (isDesktopClient) {
-    switch (router.query.theme) {
-      case "theme-dark":
-        style = `
-          ::-webkit-scrollbar {
-            border-radius: 3px !important;
-            width: 8px !important;
-            background: transparent !important;
-          }
-          
-          ::-webkit-scrollbar:hover {
-            border-radius: 3px !important;
-            width: 8px !important;
-            background: #404040 !important;
-          }
-
-          ::-webkit-scrollbar-thumb {
-            border-radius: 3px !important;
-            background: #404040 !important;
-          }
-
-          ::-webkit-scrollbar-thumb:hover {
-            border-radius: 3px !important;
-            background: #606060 !important;
-          }
-        `;
-        break;
-      case "theme-contrast-dark":
-        style = `
-          ::-webkit-scrollbar {
-            border-radius: 3px !important;
-            width: 8px !important;
-            background: transparent !important;
-          }
-
-          ::-webkit-scrollbar:hover {
-            border-radius: 3px !important;
-            width: 8px !important;
-            background: #252525 !important;
-          }
-            
-          ::-webkit-scrollbar-thumb {
-            border-radius: 3px !important;
-            background: #404040 !important;
-          }
-
-          ::-webkit-scrollbar-thumb:hover {
-            border-radius: 3px !important;
-            background: #666666 !important;
-          }
-        `;
-        break;
-      default:
-        style = `
-          ::-webkit-scrollbar {
-            border-radius: 3px !important;
-            width: 8px !important;
-            background: transparent !important;
-          }
-
-          ::-webkit-scrollbar:hover {
-            border-radius: 3px !important;
-            width: 8px !important;
-            background: #F7F7F7 !important;
-          }
-            
-          ::-webkit-scrollbar-thumb {
-            border-radius: 3px !important;
-            background: #E0E0E0 !important;
-          }
-
-          ::-webkit-scrollbar-thumb:hover {
-            border-radius: 3px !important;
-            background: #CCCCCC !important;
-          }
-        `;
-    }
-  }
-
+const MainHead = ({ title = null, description = null }) => {
   return (
     <>
       <Head>
@@ -103,22 +20,13 @@ const MainHead = ({ title = null, description = null, isDesktopClient }) => {
         <meta id="ctl00_MetaKeywords" name="keywords" content={title} />
         <meta name="description" content={description} />
         <meta name="google" content="notranslate" />
-
-        {isDesktopClient ? (
-          <style type="text/css">
-            {style}
-          </style>
-        ) : (
-          <>
-            <link rel="shortcut icon" sizes="16x16" href="/images/logo/favicons/favicon.png" type="image/png" />
-            <link rel="shortcut icon" sizes="32x32" href="/images/logo/favicons/favicon32.png" type="image/png" />
-            <link rel="shortcut icon" sizes="64x64" href="/images/logo/favicons/favicon64.png" type="image/png" />
-            <link rel="icon" sizes="96x96" href="/images/logo/favicons/favicon.ico" type="image/x-icon" />
-            <link rel="apple-touch-icon" sizes="150x150" href="/images/logo/favicons/apple150.png" type="image/png" />
-            <link rel="apple-touch-icon" sizes="310x310" href="/images/logo/favicons/apple310.png" type="image/png" />
-            <link rel="apple-touch-icon" sizes="325x325" href="/images/logo/favicons/apple325.png" type="image/png" />
-          </>
-        )}
+        <link rel="shortcut icon" sizes="16x16" href="/images/logo/favicons/favicon.png" type="image/png" />
+        <link rel="shortcut icon" sizes="32x32" href="/images/logo/favicons/favicon32.png" type="image/png" />
+        <link rel="shortcut icon" sizes="64x64" href="/images/logo/favicons/favicon64.png" type="image/png" />
+        <link rel="icon" sizes="96x96" href="/images/logo/favicons/favicon.ico" type="image/x-icon" />
+        <link rel="apple-touch-icon" sizes="150x150" href="/images/logo/favicons/apple150.png" type="image/png" />
+        <link rel="apple-touch-icon" sizes="310x310" href="/images/logo/favicons/apple310.png" type="image/png" />
+        <link rel="apple-touch-icon" sizes="325x325" href="/images/logo/favicons/apple325.png" type="image/png" />
 
         {languages.map((lng) => {
           const { key, shortKey } = lng;
@@ -133,21 +41,19 @@ const MainHead = ({ title = null, description = null, isDesktopClient }) => {
         })}
       </Head>
 
-      {!isDesktopClient &&
-        <Script
-          id="googletagmanager"
-          strategy="afterInteractive"
-          dangerouslySetInnerHTML={{
-            __html: `
+      <Script
+        id="googletagmanager"
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `
           (function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
           new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
           j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
           'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
           })(window,document,'script','dataLayer','GTM-5NW47TX');
         `,
-          }}
-        />
-      }
+        }}
+      />
     </>
   );
 };